+++ src/serialization/__serialize_311.cpp
@@ -28,8 +28,8 @@ namespace fertilized {
     if (always_register ||
         serialization_library_version >= 0) {
       ar.template register_type<TwoSideClassificationThresholdOptimizer<
-              uint8_t,
-              float,
+              int,
+              int,
               uint
 	  >>();
     }
@@ -44,18 +44,18 @@ namespace fertilized {
       const unsigned int &serialization_library_version);
 
     TemplateFuncExport DllExport std::string serialize(const TwoSideClassificationThresholdOptimizer<
-              uint8_t,
-              float,
+              int,
+              int,
               uint
 	  > *, const bool &);
     TemplateFuncExport DllExport TwoSideClassificationThresholdOptimizer<
-              uint8_t,
-              float,
+              int,
+              int,
               uint
 	  >* deserialize(std::stringstream &);
     TemplateFuncExport DllExport void deserialize(std::stringstream &, TwoSideClassificationThresholdOptimizer<
-              uint8_t,
-              float,
+              int,
+              int,
               uint
 	  >*);
 }  // namespace fertilized
@@ -67,8 +67,8 @@ namespace serialization {
 
 template <>
 struct version<TwoSideClassificationThresholdOptimizer<
-              uint8_t,
-              float,
+              int,
+              int,
               uint
 	  >> {
     typedef mpl::int_<FERTILIZED_VERSION_COUNT> type;
