+++ src/serialization/__serialize_317.cpp
@@ -28,9 +28,9 @@ namespace fertilized {
     if (always_register ||
         serialization_library_version >= 0) {
       ar.template register_type<VarianceClassificationThresholdOptimizer<
-              float,
-              float,
-              int16_t
+              double,
+              double,
+              uint
 	  >>();
     }
   };
@@ -44,19 +44,19 @@ namespace fertilized {
       const unsigned int &serialization_library_version);
 
     TemplateFuncExport DllExport std::string serialize(const VarianceClassificationThresholdOptimizer<
-              float,
-              float,
-              int16_t
+              double,
+              double,
+              uint
 	  > *, const bool &);
     TemplateFuncExport DllExport VarianceClassificationThresholdOptimizer<
-              float,
-              float,
-              int16_t
+              double,
+              double,
+              uint
 	  >* deserialize(std::stringstream &);
     TemplateFuncExport DllExport void deserialize(std::stringstream &, VarianceClassificationThresholdOptimizer<
-              float,
-              float,
-              int16_t
+              double,
+              double,
+              uint
 	  >*);
 }  // namespace fertilized
 
@@ -67,9 +67,9 @@ namespace serialization {
 
 template <>
 struct version<VarianceClassificationThresholdOptimizer<
-              float,
-              float,
-              int16_t
+              double,
+              double,
+              uint
 	  >> {
     typedef mpl::int_<FERTILIZED_VERSION_COUNT> type;
     typedef mpl::integral_c_tag tag;
