+++ src/serialization/__serialize_345.cpp
@@ -28,8 +28,8 @@ namespace fertilized {
     if (always_register ||
         serialization_library_version >= 101) {
       ar.template register_type<BoostingLeafManager<
-              uint8_t,
-              int16_t
+              int,
+              uint
 	  >>();
     }
   };
@@ -43,16 +43,16 @@ namespace fertilized {
       const unsigned int &serialization_library_version);
 
     TemplateFuncExport DllExport std::string serialize(const BoostingLeafManager<
-              uint8_t,
-              int16_t
+              int,
+              uint
 	  > *, const bool &);
     TemplateFuncExport DllExport BoostingLeafManager<
-              uint8_t,
-              int16_t
+              int,
+              uint
 	  >* deserialize(std::stringstream &);
     TemplateFuncExport DllExport void deserialize(std::stringstream &, BoostingLeafManager<
-              uint8_t,
-              int16_t
+              int,
+              uint
 	  >*);
 }  // namespace fertilized
 
@@ -63,8 +63,8 @@ namespace serialization {
 
 template <>
 struct version<BoostingLeafManager<
-              uint8_t,
-              int16_t
+              int,
+              uint
 	  >> {
     typedef mpl::int_<FERTILIZED_VERSION_COUNT> type;
     typedef mpl::integral_c_tag tag;
